pyproject.toml
[project]
name = "smartrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepeval>=3.9.7",
    "fastapi>=0.136.0",
    "google-cloud-storage>=3.10.1",
    "httpx>=0.28.1",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.14",
    "langchain-qdrant>=1.1.0",
    "langchain-text-splitters>=1.1.2",
    "langsmith>=0.7.32",
    "motor>=3.7.1",
    "pydantic>=2.13.2",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "requests>=2.33.1",
    "streamlit>=1.56.0",
    "tiktoken>=0.12.0",
    "unstructured>=0.22.21",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "mypy>=1.20.1",
    "ruff>=0.15.11",
]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
# E: pycodestyle errors, F: Pyflakes, I: isort (auto-sorts imports)
select = ["E", "F", "I"]
ignore = []

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true
warn_unused_ignores = true
explicit_package_bases = true
